Topology used
Green nodes refer to TCP sources and sinkers while the yellow ones are UDP sources and sinkers.
All the nodes are connected via peer-to-peer architecture. Nodes have a bandwidth of 50Mbps and delay of 10ms. The blue boxes are the three routers R1, R2, R3. R1-R2 have a bandwidth of 20Mbps and delay of 20ms while R2-R3 have a bandwidth of 10Mbps and delay of 10ms.
Number of packets sent = 100000
Size of each packet = 1040 bytes
Simulation Time = 200ms